% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_blocks.R
\name{associated_block}
\alias{associated_block}
\title{Find an associated block}
\usage{
associated_block(file, line)
}
\arguments{
\item{file}{(\code{character[1]}) A file path associated with a tag}

\item{line}{(\code{integer[1]}) A line number in the file where the tag was found}
}
\value{
(\code{\link[roxygen2:roxy_block]{roxygen2::roxy_block()}} | NULL) A \code{roxy_block} if one could be associated, or \code{NULL} if not.
}
\description{
Given a file and line, discover which block the tag is associated with
}
\keyword{internal}
