% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cli.R
\name{convert_helpers}
\alias{convert_helpers}
\alias{preview_convert_edits}
\alias{preview_convert_edit}
\title{Conversion Helpers}
\usage{
preview_convert_edits(edits, n = 1)

preview_convert_edit(edit)
}
\arguments{
\item{n}{The number of tags to preview}

\item{format}{(\code{string}) A regular expression to use to match existing param definitions.}

\item{unmatched}{(\code{logical[1]}) Indicates whether \verb{@param} tags that don't match the expected pattern
should be converted anyways. In such cases, their entire description is
migrated to a new \verb{@typed} description and the new \verb{@typed} tag's type is
left blank.}

\item{tags}{A list of \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}s to preview}
}
\value{
\code{NULL}
}
\description{
Various functions for supporting conversion from standard roxygen tags to
\verb{@typed} tags.
}
\section{Functions}{
\itemize{
\item \code{preview_convert_edits()}: Preview diffs after applying conversion rules

\item \code{preview_convert_edit()}: Preview diffs after applying conversion rules

}}
