% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_match_format}
\alias{convert_match_format}
\title{Match a conversion format and structure results}
\usage{
convert_match_format(x, format)
}
\arguments{
\item{x}{(\code{character[1]}) Content to match.}

\item{format}{(\code{character[1]}) A regular expression, optionally containing named capture groups for \code{type}
and \code{description}, which will be used for restructuring the tag as a
\code{roxytypes}-equivalent tag.}
}
\value{
(\verb{: list}) A named list of \code{type}, \code{description} and \code{matched} fields. \code{type} and
\code{description} represent the result of captured groups. If no capture groups
were used, the raw string is used as a description. \code{matched} is a
\code{logical[1]} indicating whether the provided format matched against the
provided input.
}
\description{
Match a conversion format and structure results
}
\seealso{
Other convert: 
\code{\link{build_convert_edits}()},
\code{\link{convert_tag}()},
\code{\link{make_convert_edits}()},
\code{\link{tag_edit}()}
}
\concept{convert}
\keyword{internal}
