% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_params}
\alias{convert_params}
\title{Convert \verb{@param} definitions to \verb{@typed} parameters}
\usage{
convert_params(
  format,
  type = re_backticked(),
  description = re_any(),
  unmatched = FALSE,
  ...,
  path = ".",
  format_re,
  verbose = interactive()
)
}
\arguments{
\item{format}{An optional \code{glue}-style format to use to parse types and
descriptions for conversion to \verb{@typed} tags. Available \code{glue} keywords
include \code{type} and \code{description}, as well as any ellipsis arguments.}

\item{type}{(\code{character[1]}) A regular expression to use for matching a \code{{type}}
within the format string. Defaults to matching content between backticks.}

\item{description}{(\code{character[1]}) A regular expression to use for matching a
\code{{description}} within the format string.}

\item{...}{Additional capture groups (unused)}

\item{path}{(\code{character[1]}) A path in which to look for \code{roxygen2} documented R scripts.}

\item{verbose}{(\code{logical[1]}) Whether to use a command line interface to interactively step
through changes.}
}
\description{
Convert \verb{@param} definitions to \verb{@typed} parameters
}
\examples{
convert_params("(`{type}`) {description}")

}
