% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{default_format}
\alias{default_format}
\title{Default formatter for \verb{@typed}}
\usage{
default_format(x, name, type, default = NULL, description, ...)
}
\arguments{
\item{x}{([roxygen2::roxy_tag()]) The tag to format.}

\item{name, type, default, description}{(`character(1)`; Default = `NULL`) Fields parsed from the \verb{@typed} tag.}

\item{...}{Additional arguments unused.}
}
\value{
A formatted character value.
}
\description{
Adds special cases for when the type uses other roxygen2 syntax
}
\keyword{internal}
