% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{make_convert_edits}
\alias{make_convert_edits}
\title{Make tag edits}
\usage{
make_convert_edits(edits)
}
\arguments{
\item{edits}{(\code{data.frame}) A collection of edits (one edit per row), as produced by
\code{\link[=tag_edit]{tag_edit()}}.}
}
\value{
(\code{integer[1]}) The number of edits that were made.
}
\description{
Make tag edits
}
\seealso{
Other convert: 
\code{\link{build_convert_edits}()},
\code{\link{convert_match_format}()},
\code{\link{convert_tag}()},
\code{\link{tag_edit}()}
}
\concept{convert}
\keyword{internal}
