% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dcf_asis}
\alias{read_dcf_asis}
\title{A helper to reliably read DCF files}
\usage{
read_dcf_asis(path)
}
\arguments{
\item{path}{(\code{character[1]}) A file path to a DESCRIPTION file.}
}
\value{
(\code{data.frame}) The result of \code{\link[=read.dcf]{read.dcf()}}.
}
\description{
A helper to reliably read DCF files
}
