% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_capture}
\alias{regex_capture}
\title{Capture regex groups}
\usage{
regex_capture(pattern, x, ...)
}
\arguments{
\item{pattern}{(\code{character[1]}) A regex pattern to use for capturing groups.}

\item{x}{(\code{character}) A vector of strings to match against.}

\item{...}{Additional arguments passed to \code{\link[=regexpr]{regexpr()}}.}
}
\description{
Captures regex groups and returns a named matrix of groups with one column
per capture group and one row per element in \code{x}.
}
\keyword{internal}
