% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{regex_helpers}
\alias{regex_helpers}
\alias{re_backticked}
\alias{re_any}
\title{Regular expression helpers}
\usage{
re_backticked()

re_any()
}
\description{
Regular expression helpers
}
\section{Functions}{
\itemize{
\item \code{re_backticked()}: Match within backticks

\item \code{re_any()}: Match any

}}
\examples{
text <- "@param (`test(\")\")`)"
pattern <- sprintf("`\%s`", re_backticked())
m <- regexec(pattern, text, perl = TRUE)
regmatches(text, m)[[1]]
# [1] "`test(\")\")`"

}
