% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{roxy_tag_parse.roxy_tag_typed}
\alias{roxy_tag_parse.roxy_tag_typed}
\title{\code{roxygen2} \verb{@typed} tag parsing}
\usage{
\method{roxy_tag_parse}{roxy_tag_typed}(x)
}
\arguments{
\item{x}{A tag}
}
\description{
Parse a \verb{@typed} tag and return parsed components as value
}
