% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typedreturn.R
\name{roxy_tag_parse.roxy_tag_typedreturn}
\alias{roxy_tag_parse.roxy_tag_typedreturn}
\title{\code{roxygen2} \verb{@typedreturn} tag parsing}
\usage{
\method{roxy_tag_parse}{roxy_tag_typedreturn}(x)
}
\arguments{
\item{x}{A tag}
}
\value{
(\verb{roxygen2 tag}) A parsed \code{roxygen2} \verb{@typedreturn} rd_tag.
}
\description{
Parse a \verb{@typedreturn} tag and return parsed components as value
}
