% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_blocks.R
\name{roxygen_blocks}
\alias{roxygen_blocks}
\title{Fetch \code{roxygen2} blocks}
\usage{
roxygen_blocks(path = getwd(), refresh = FALSE, cache = TRUE)
}
\description{
Avoid recomputing \code{roxygen2}s parsing by saving the blocks after the first
tag is hit.
}
\keyword{internal}
