% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_and_trim}
\alias{split_and_trim}
\title{Split and trim a string}
\usage{
split_and_trim(x)
}
\arguments{
\item{x}{(\code{character[1]}) A string to split into lines and trim.}
}
\value{
(\code{x: character}) A character vector of trimed lines.
}
\description{
Split and trim a string
}
\keyword{internal}
