% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{tag_edit}
\alias{tag_edit}
\title{Built a conversion edit}
\usage{
tag_edit(tag, new, matched)
}
\arguments{
\item{tag}{(\code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}) A \code{roxygen2} tag to convert.}

\item{new}{(\code{character}) The new content used to replace the tag.}

\item{matched}{(\code{logical[1]}) Whether the content was generated based on a match of a specified format.}
}
\value{
(\code{data.frame}) A single-observation dataset representing information for a tag edit. The
\code{data.frame} row includes variables:
\itemize{
\item \code{file}: (\code{character[1]}) The source file for the tag.
\item \code{line}: (\code{integer[1]}) The first line of the tag.
\item \code{n}: (\code{integer[1]}) The number of lines the tag spans.
\item \code{matched}: (\code{logical[1]}) Whether the tag matched a specified format.
\item \code{new}: (\code{list[1](character)}) The new contents to replace the tag.
}
}
\description{
Built a conversion edit
}
\seealso{
Other convert: 
\code{\link{build_convert_edits}()},
\code{\link{convert_match_format}()},
\code{\link{convert_tag}()},
\code{\link{make_convert_edits}()}
}
\concept{convert}
\keyword{internal}
