% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{tags}
\alias{tags}
\alias{default}
\alias{typed}
\title{\code{roxytypes} tags}
\usage{
#' @typed <var>: <type>
#'   <description>

#' @typed <var>: <type> @default <default>
#'   <description>
}
\description{
The \verb{@typed} tag introduces a syntax for defining parameter types as a
\code{roxygen2} tag.
}
\details{
There are a couple important syntactic features:
\itemize{
\item \code{:} delimiter between the variable name and type.
\item \verb{\\n} after the type to separate it from the description.
\item \verb{@default} tag can be used to optionally specify a default parameter
value. If not set, and \code{derive_defaults} is enabled, then atomic,
length-1 parameter defaults will automatically be used to populate
default arguments in documentation.
}
}
