% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{try_parse_typed}
\alias{try_parse_typed}
\title{Parse a raw \verb{@typed} tag}
\usage{
try_parse_typed(raw)
}
\arguments{
\item{raw}{(\code{character[1]}) A raw string from which to try to parse typed parameter definitions.}
}
\description{
Parse a raw \verb{@typed} tag
}
\keyword{internal}
