% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typedreturn.R
\name{try_parse_typedreturn}
\alias{try_parse_typedreturn}
\title{Parse a raw \verb{@typedreturn} tag}
\usage{
try_parse_typedreturn(raw)
}
\arguments{
\item{raw}{(\code{character[1]}) A raw string from which to try to parse typed parameter definitions.}
}
\description{
Parse a raw \verb{@typedreturn} tag
}
\keyword{internal}
