% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vapply_helpers}
\alias{vapply_helpers}
\alias{vlapply}
\alias{vcapply}
\title{vapply helpers}
\usage{
vlapply(..., FUN.VALUE = logical(1L))

vcapply(..., FUN.VALUE = character(1L))
}
\arguments{
\item{...}{Passed to \code{\link[=vapply]{vapply()}}.}

\item{FUN.VALUE}{A prototype signature to use for \code{\link[=vapply]{vapply()}}.}
}
\description{
vapply helpers
}
\section{Functions}{
\itemize{
\item \code{vlapply()}: Logical vapply

\item \code{vcapply()}: Character vapply

}}
\keyword{internal}
