% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_roxy_field_subclass}
\alias{with_roxy_field_subclass}
\title{A helper to apply field names to all roxy_tag val fields}
\usage{
with_roxy_field_subclass(x)
}
\arguments{
\item{x}{(\code{list}) A named list of tag val contents}
}
\value{
(\verb{: list}) A nearly identical list, where elements have additional subclasses based
on their field names.
}
\description{
A helper to apply field names to all roxy_tag val fields
}
