% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_config.R
\name{config}
\alias{config}
\title{\code{roxytypes} Config}
\usage{
config(path = getwd(), refresh = FALSE, cache = TRUE)
}
\value{
(`list`) A named list of configured behaviors.
}
\description{
\code{roxytypes} exposes a few configuration options for helping to fine-tune your
documentation. These are stored as key-values in a \code{list} in either the
\code{Config/roxytypes} field of your \code{DESCRIPTION} file, or in a
\code{./man/roxytypes/meta.R} file within your package.
}
\details{
The available settings are listed below. Some fields are nested, which are
shown by concatenating nested keys using \code{$}.
\itemize{
\item \code{format}: An optional \code{glue}-style string, which can assume values for
\code{name}, \code{type}, \code{default} and \code{description}. See \code{?roxytypes::tags} for
details on the source of each of these strings.
\item \code{defaults$derive}: Whether to try to derive default values for
documentation. When a function's formal arguments are length-1 atomic
values (or \code{NULL}), they can be derived for your documentation.
\item \code{defaults$missing}: Either \code{NULL} or a character value. If a character
value is provided, it is used in documentation as the default value when
no default is defined. If \code{NULL}, missing default values are not included
in documentation.
\item \code{defaults$warn_undocumented}: If \code{TRUE}, alerts are raised if defaults are
defined, but not documented.
\item \code{verbose}: If \code{TRUE}, emit extra diagnostic alerts while processing the
package.
}
}
\keyword{internal}
