% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_config.R
\name{config_helpers}
\alias{config_helpers}
\alias{config_find_from}
\alias{config_from_desc}
\alias{config_from_file}
\title{Configuration}
\usage{
config_find_from(path = ".")

config_from_desc(path = ".")

config_from_file(path = ".")
}
\arguments{
\item{path}{(\code{character[1]}) A file path to use when searching for a config file. Either the file path
to a \code{DESCRIPTION} or the root path of a package, depending on the context
of the function.}
}
\description{
Various functions for loading, caching and performing configured behaviors
using a user-supplied configuration file.
}
\details{
This constant is used as a variable name in the package environment while
documentation is being built to avoid constantly parsing configurations
during evaluation of each tag.
}
\section{Functions}{
\itemize{
\item \code{config_find_from()}: Load a configuration from a path

\item \code{config_from_desc()}: Load a configuration from a DESCRIPTION file

\item \code{config_from_file()}: Load a configuration from a dotfile

}}
\keyword{internal}
