% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_helpers}
\alias{convert_helpers}
\alias{convert_edit_params}
\alias{convert_cli}
\alias{convert_param}
\alias{preview_convert_diffs}
\alias{preview_convert_diff}
\title{Conversion Helpers}
\usage{
convert_edit_params(format, tags, unmatched = FALSE)

convert_cli(format, tags)

convert_param(format, param)

preview_convert_diffs(format, tags, n = 1)

preview_convert_diff(format, tag)
}
\arguments{
\item{format}{(\code{string}) A regular expression to use to match existing param definitions.}

\item{tags}{A list of \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}s to preview}

\item{unmatched}{(\code{logical[1]}) Indicates whether \verb{@param} tags that don't match the expected pattern
should be converted anyways. In such cases, their entire description is
migrated to a new \verb{@typed} description and the new \verb{@typed} tag's type is
left blank.}

\item{n}{The number of tags to preview}
}
\value{
\code{TRUE} if proceeding with changes, or \code{FALSE} if aborting changes

\code{NULL}
}
\description{
Various functions for supporting conversion from standard roxygen tags to
\verb{@typed} tags.
}
\section{Functions}{
\itemize{
\item \code{convert_edit_params()}: Commit edits to \verb{@param} tags

\item \code{convert_cli()}: Display a CLI to preview converted tags before applying changes

\item \code{convert_param()}: Convert a single tag to a \verb{@typed} tag

Attempt to match the existing \verb{@param} description and use the matched
components to populate a new \verb{@typed} tag.

\item \code{preview_convert_diffs()}: Preview diffs after applying conversion rules

\item \code{preview_convert_diff()}: Preview diffs after applying conversion rules

}}
