% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{default_format}
\alias{default_format}
\title{Default formatter for \verb{@typed}}
\usage{
default_format(x, name, type, description, ...)
}
\arguments{
\item{x}{(\code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}) The tag to format.}

\item{name, type, description}{(\code{character(1)}) Fields parsed from the \verb{@typed} tag.}

\item{...}{Additional arguments unused.}
}
\value{
A formatted character value.
}
\description{
Adds special cases for when the type uses other roxygen2 syntax
}
