% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typedreturn.R
\name{roxy_tag_rd.roxy_tag_typedreturn}
\alias{roxy_tag_rd.roxy_tag_typedreturn}
\title{\code{roxygen2} \verb{@typedreturn} tag rd section population}
\usage{
\method{roxy_tag_rd}{roxy_tag_typedreturn}(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
(\link[roxygen2:rd_section]{roxygen2::rd_section}) A \code{roxygen2} \verb{@value} rd_tag with formatted type information.
}
\description{
Push typed fields into \verb{@param} section
}
