% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typed.R
\name{tags}
\alias{tags}
\alias{typed}
\title{\code{roxytypes} tags}
\usage{
#' @typed <var>: <type>
#'   <description>
}
\description{
The \verb{@typed} tag introduces a syntax for defining parameter types as a
\code{roxygen2} tag.
}
\details{
Be aware that there are a few syntactic requirements:
\itemize{
\item \code{:} delimiter between the variable name and type.
\item \verb{\\n} after the type to separate it from the description.
}
}
