% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_typedreturn.R
\name{typedreturn}
\alias{typedreturn}
\title{\code{roxygen2} \verb{@typedreturn} tag}
\usage{
#' @typedreturn <type>
#'   <description>
}
\description{
The \verb{@typedreturn} tag introduces a syntax for defining return types as a
\code{roxygen2} tag.
}
\details{
There is one important syntactic features:
\itemize{
\item \verb{\\n} after the type to separate it from the description.
}
}
